requirements.txt
#Chocolatey and 
absl-py                 ==2.1.0
aiohappyeyeballs        ==2.4.3
aiohttp                 ==3.11.7
aiosignal               ==1.3.1
annotated-types         ==0.7.0
anyascii                ==0.3.2
anyio                   ==4.6.2.post1
asttokens               ==2.4.1
async-timeout           ==5.0.1
attrs                   ==v24.2.0
audioread               ==3.0.1
babel                   ==2.16.0
bangla                  ==0.0.2
blinker                 ==1.9.0
#blis                    ==1.0.1
bnnumerizer             ==0.0.2
bnunicodenormalizer     ==0.1.7
catalogue               ==2.0.10
certifi                 ==2024.8.30
cffi                    ==1.17.1
charset-normalizer      ==3.4.0
click                   ==8.1.7
cloudpathlib            ==0.20.0
colorama                ==0.4.6
confection              ==0.1.5
#contourpy               ==1.3.1
coqpit                  ==0.0.17
customtkinter           ==5.2.2
cycler                  ==0.12.1
cymem                   ==2.0.8
Cython                  ==3.0.11
darkdetect              ==0.8.0
dateparser              ==1.1.8
decorator               ==5.1.1
distro                  ==1.9.0
docopt                  ==0.6.2
einops                  ==0.8.0
encodec                 ==0.1.1
exceptiongroup          ==1.2.2
executing               ==2.1.0
ffmpeg                  ==1.3
filelock                ==3.16.1
Flask                   ==3.1.0
fonttools               ==4.55.0
frozenlist              ==1.5.0
fsspec                  ==2024.10.0
g2pkk                   ==0.1.2
groq                    ==0.12.0
grpcio                  ==1.68.0
#gruut                   ==2.2.3
#gruut-ipa               ==0.13.0
#gruut_lang_de           ==2.0.1
#gruut_lang_en           ==2.0.1
#gruut_lang_es           ==2.0.1
#gruut_lang_fr           ==2.0.2
h11                     ==0.14.0
hangul-romanize         ==0.1.0
httpcore                ==1.0.7
httpx                   ==0.27.2
huggingface-hub         ==0.26.2
idna                    ==3.10
inflect                 ==7.4.0
ipython                 ==8.29.0
itsdangerous            ==2.2.0
jamo                    ==0.4.1
jedi                    ==0.19.2
jieba                   ==0.42.1
Jinja2                  ==3.1.4
joblib                  ==1.4.2
jsonlines               ==1.2.0
kiwisolver              ==1.4.7
langcodes               ==3.5.0
language_data           ==1.3.0
lazy_loader             ==0.4
#librosa                 ==0.10.2.post1
llvmlite                ==0.43.0
marisa-trie             ==1.2.1
Markdown                ==3.7
markdown-it-py          ==3.0.0
MarkupSafe              ==3.0.2
matplotlib              ==3.9.2
matplotlib-inline       ==0.1.7
mdurl                   ==0.1.2
more-itertools          ==10.5.0
mpmath                  ==1.3.0
msgpack                 ==1.1.0
multidict               ==6.1.0
murmurhash              ==1.0.10
mutagen                 ==1.47.0
networkx                ==2.8.8
nltk                    ==3.9.1
num2words               ==0.5.13
numba                   ==0.60.0
#numpy                   ==1.26.4
openai-whisper          ==20240930
packaging               ==24.2
pandas                  ==1.5.3
parso                   ==0.8.4
pillow                  ==11.0.0
pip                     ==24.3.1
platformdirs            ==4.3.6
pooch                   ==1.8.2
preshed                 ==3.0.9
prompt_toolkit          ==3.0.48
propcache               ==0.2.0
protobuf                ==5.28.3
psutil                  ==6.1.0
pure_eval               ==0.2.3
pycparser               ==2.22
pydantic                ==2.10.1
pydantic_core           ==2.27.1
pydub                   ==0.25.1
Pygments                ==2.18.0
pynndescent             ==0.5.13
pyparsing               ==3.2.0
pypinyin                ==0.53.0
pysbd                   ==0.3.4
python-crfsuite         ==0.9.11
python-dateutil         ==2.9.0.post0
pytz                    ==2024.2
PyYAML                  ==6.0.2
regex                   ==2024.11.6
requests                ==2.32.3
rich                    ==13.9.4
safetensors             ==0.4.5
#scikit-learn            ==1.5.2
#scipy                   ==1.14.1
setuptools              ==65.5.0
shellingham             ==1.5.4
six                     ==1.16.0
smart-open              ==7.0.5
sniffio                 ==1.3.1
soundfile               ==0.12.1
#soxr                    ==0.5.0.post1
#spacy                   ==3.8.2
spacy-legacy            ==3.0.12
spacy-loggers           ==1.0.5
srsly                   ==2.4.8
stack-data              ==0.6.3
SudachiDict-core        ==20241021
SudachiPy               ==0.6.9
sympy                   ==1.13.1
#tensorboard             ==2.18.0
tensorboard-data-server ==0.7.2
#thinc                   ==8.3.2
threadpoolctl           ==3.5.0
tiktoken                ==0.8.0
tokenizers              ==0.20.3
#torch                   ==2.5.1+cu121
#torchaudio              ==2.5.1+cu121
#torchvision             ==0.20.1+cu121
tqdm                    ==4.67.0
trainer                 ==0.0.36
traitlets               ==5.14.3
transformers            ==4.46.3
#TTS                     ==0.22.0
typeguard               ==4.4.1
typer                   ==0.13.1
typing_extensions       ==4.12.2
tzdata                  ==2024.2
tzlocal                 ==5.2
umap-learn              ==0.5.7
Unidecode               ==1.3.8
urllib3                 ==2.2.3
wasabi                  ==1.1.3
wcwidth                 ==0.2.13
weasel                  ==0.4.1
Werkzeug                ==3.1.3
wrapt                   ==1.17.0
yarl                    ==1.18.0